+++ src/main.c
@@ -2,8 +2,8 @@
 
 int main() {
   canvas_t *c = canvas_create(20, 10);
-  for (int i = 0; i < c->height; i++) {
-    for (int j = 0; j < c->width; j++) {
+  for (size_t i = 0; i < c->height; i++) {
+    for (size_t j = 0; j < c->width; j++) {
       canvas_set_pixel(c, j, i, '#');
     }
   }
